requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
jinja2
python-multipart
